Pushed and popped scenes after all updated
When fluently navigated two scenes pushed two scenes but in between the didShow of the first push fired. It looked like a back navigation because the viewControllers < crumbs so it navigated back! Then the second push happened then the setState from the navigateBack which meant it popped the second push!
Applied pushes and pops together after all subview updates which prevents the first push firing the didShow