Added locking mechanism for ios back
To simulate slow down a render but putting in a 2 sec sleep.
Then push A,  push B, push C - before C displays press ios back button. Without the locking this would first go back to A (correct) and then display C (wrong) and then go back to A again! That's because the push to C comes back after the ios back. The ios back pop is on the client so happens instantly and then when the C push returns it displays C and then the setState for the pop returns and goes back to A. The locking mechanism is same as in place on TextInput to prevent the typed text and setState(text) getting out of synch