Finished Activity when SceneView removed
Problem is that removing the SceneView removes all it's children and this makes the Activity go blank as it animates away!!! Overrode removeAllViews of SceneManager so children are kept. Faked the ChildCount of SceneManager to 0 so that React doesn't loop through it's children removing all descendants! This needs revisiting because it prevents the NativeViewHierarchyManager from clearing the Tag arrays. Don't know how to get around this!