Received new intent on back navigation
Tested in current navigation router and it's an existing bug. If you go from A --> B --> C --> D --> E and then navigate back 3 you end up back at D and not B!!! That's because B and D are both SceneActivity with singleTop and so, although it goes back to B it uses the intent from D. I think anyway. It's confusing. I think it reuses the Activity of D at position B!? I guess it makes a bit of sense if navigating forward but navigating back!? Put onNewIntent to read B's crumb and load in B's view after removing D's view!!