Changed fluent navigation then back check
if navigate from A --> B --> C --> D --> E and fluently from C to E, then navigate back 3 to B then it will come in the create activity of D because it wasn't created. At that time it will have the crumb for D but there won't be a scene view for it. Still created the host view because onNewIntent is about to run and needs a container