Calculated enter/exit anim when scene removed
Couldn't read the enter and exit in removeView because it's called before they're set on the current scene!! So moved the start and finish activities into onAfterUpdateTransaction. This is called after views added/removed and props set. But onAfterUpdateTransaction is only called if the props of the NavigationStackView have changed. So forced it to change every navigation by setting url and oldUrl props