Ensured React Native clears tags cache
When a scene's removed trick React Native in thinking it has no children so that it they don't disappear when the scene animates away. The problem with this is that the dropView of NativeViewHierarchyManager doesn't clear it's tag cache for these children. The caches just grow in size when navigating around. So created a SceneBin that says it's children are the discarded scenes  children. That way when the SceneBin is removed from the NavigationStack then these children are dropped and the tag caches cleared! For the SceneBin to be added and removed by updating it's React key whenever 2 seconds have passed between renders