Prevented error when double tap back
The locking mechanism should handle this scenario but the didShowViewController fires late (unlike textChange on TextInput). So it's possible to double tap so fast that didShowViewController doesn't fire for the first tap - which means the nativeEventCoutn doesn't increment and locking mechanism doesn't work. A --> B --> C --> D and tap back 3 times fast. The first back fires didShowViewController and starts a setState. Before it returns the second two taps come in so fast that they don't fire didShowViewController. Then the first setState returns, the counts match and so it wants to do a forward navigation to C - this throws error because C already has a parent view so can't have another