Fixed shared els when navigating back
The Android intent to actiivity look up is a nightmare. When navigateUpTo and pass intent it won't find the activity that was created for that intent!!! It finds any activity with a matching type!? A --> B --> C --> D and navigate back 2 will use the activity for C because both have SceneActivtiy type and C is latest. This isn't a problem normally because onNewIntent finds the right SceneView and adds it as a child. The problem happens with shared elements transitions. If there was a shared el transition when going from A to B, then want it to run when going back from B to A. But the actual activity is C (not B) because it's been reused. So if call finishAfterTransition current activity it will go back to B not A!?! craziness.
The only way I could find around it is to ensure all activities have a different type. This guarantees that navigateUpTo always finds the right activity which means shared element navigation works. The catch is that can't there's a finite number of activity types (created 20). So disabled shared elements for back navigation from crumb 21 and beyond.