Hid navigation bar after setting props
Problem with doing it in didMoveToWindow is that it's called many times, not just first time added. It's called when navigation to next screen, causing the bar to be hidden temporarily. Wouldn't need to do this is the hidden property was on the viewController's navigationItem but it's on the shared navigationController