Passed the current context to navigateLink
On Mobile and Native, the navigation was only idempotent for the next and previous state and data but not for the old state and data. It was ok to double tapping because it wouldn't change the crumb trail. But double tapping would change the old state. For example, a double tap a link that navigates A --> B. The first tap the old State is A. The second tap the old State is B because it took the old State from the top level context.
Instead, taken the context from the async state navigator not the top level state navigator. That way the old State is always A because the async state navigator stays  the same for scene A