Started new activity for first scene
If add directly to MainActivity then can't replace the first scene because not safe to finish MainActivity. If finish MainActivity then can't replace with MainActivity because no way to set the content; and can't repalce with SceneActivity because React Native relies on there being a MainActivity, e.g., for dev menu.
Had to manually call onAfterUpdateTransaction because, on first load, it's called before the views have been added?! Called it when SceneBin added because that's the last child