Changed scene key notation
Start with A -> B -> C
Fluently navigate to A -> D -> C
Navigate back to A -> D
Fluently navigate to A -> B
This should replace D with B, animating D out and B in. But it didn't animate. The key for the 2nd screen started as 1_B. The final fluent navigation changed it to 1_B - the same! - so it didn't animate the replace.
Changed keys. Instead of using state keys, add a '+' on the end every time it's replaced. In this case, it starts as '1'. The final fluent changes it to '1+' and the replace animates