Checked scenes updated before navigating
Don't need this on iOS because it does the check. It only calls didUpdateReactSubviews if the scenes have changed (added or removed). Java doesn't do this chekc so added a boolean and updated it whenever a scene's added or removed. Only need this now because navigation can happen when crumbs don't change (replacing)