Used prev state context to check for replace
Button that navigates from A --> B. Click it repeatedly before first navigation completes (add javascript breakpoint to pause first click). The first click sets the oldState to A and state to B. The second click does the same because navigation is idempotent. But this time the prevKeys length matches the keys length so it think it's a replace!!
Comparing prevKeys and keys so should compare preState and state.
Could change to compare oldUrl length and keys length but think this would go wrong when it is a replace because the replace would happen twice (need to check!)