Prevented reuse of replaced screens
Navigate fluently from A to A --> B to A --> C. Then press back and repeat quickly (less than 1 sec) then end up with C looking lilke B. So get B entering and B leaving!
It's because the replacee screen is sitting around still rendered (not removed until a second to give time for pop animation to complete). Then when go from A to B, then the replaced screen updates to look like B. When actually do a replace, the dormant replaced screen is revived and it still looks like B - it thinks a replace has happened so doesn't change to C.
Will eventually remove dormant screens on a timeout but this could still happen if not quick enough at removing. There's no need to reuse dormant replacee screens - they're being reused by accident. So changed to guaratnee each replacee as a unique key so can't be reused.