Removed popped scenes on 2nd pass
No need for tick. Better to makr items as popped and remove them on next navgiation. If the next navigation happens then the first popped animation has completed - safe to say. On Android that's true because it only animates two scenes, not the whole stack. On iOS there's no back button where can do multiple taps in one go, really have to wait for animation to complete to do the next navigation