Added scene context for android back
Navigate A → B → C with an android back handler in C that returns true but does nothing. Navigate back programmatically to B and then press the android back. If happesn quick enough so that scene C is still around in javascript then C's android back handler will block it going back to A.
Created a SceneContext with a canHandleBack function. It only returns true if the current scene is the top of the stack. This means that C isn't in the stack and the check returns false when called from C's back handler