Replaced scenes can't handle back
Navigate fluently from A to B and have an android back handler in A that returns true but does nothing. Pressing android back on B should close the app because it's the first scene. If A is still around then A will block the back. Even with call to canHandleBack on SceneContext it still blocks because it's crumb 0 and thinks it's top of stack.
Add a replace boolean in scene state that tracks when the scene's replaced and use this to return false in canHandleBack