Tracked top of stack
Not right having it as a function that need to call to evaluate. It should update whenever it's top of stack status changes.  This does mean that two scenes can update at once now (if both sccenes are SceneContext consumers) but it's the right way to set a context value