Rerendered on tracker change
Otherwise value isn't propagate to consumers. Does mean that old scene always renders along with new scene but price worth paying.
Showed up error with accessing state and data in render!!! If navigationEvent is set then must use state and data from that because now scene is rendered when popped and so no longer in stack!