Checked if keys change before replacing
In Zoom example, select a color and then change a color on details screen. It thinks it's a replace because the crumb == currentCrumb. But it's not a replace because the state's haven't changed. It comes in here because the sharedElements have changed but the keys haven't. So added check to see if really a replace by comarping the oldKey with current key.
Don't have the problem on iOS because iOS doesn't have shared Element (or animation) properties so React Native doesn't call didSetProps