Unmounted scenes with same pop times
A scene is recyclable if it's Activiity class isn't unique (not one of the first 21). Navigate from ...--> A --> B --> C where A, B and C all have the same Activity class. Then if navigate back 2 to A it will reuse B's Activiy! So Activity B won't be destroyed - only Activity C is destoryed. In React B and C have the same popTime because they were popped together. So can find item B from item C's popTime and unmount it when C is unmounted