Prevented out of sequence unmount
A --> B --> C --> D with B, C and D the same activity class and A different. Then press android back fast three times. It ends up unmounting scene B first so and the screen goes blank as it animates out back to A. Think that the navigation happens so fast that it puts B in C because Activity C is still around when navigating back to B. But then navigation back to A happens which destorys Activity C - so screen B is unmounted and then B goes blank?!
Anyway, slowed the navigation back down so that it nabigation back to B can't happen until screen C unmounts because C won't free up it's back handler until it unmounts