Prevented pop if scene re-parented
A --> B --> C --> D with B, C and D sharing same Activity class and A with different one.  Navigate back to and get A --> B --> B in terms of Activities because B reuses Activity C. Then navigate back 1 and /Activity for scene B is destroyed twice! The first time is the first B and the the screen goes blank because the second B is still animating!
Change so that onDestory only unmounts the scene (calls onPopped) if it's parent matches the Activities root view