Prevented double click doing a replace
Double click A --> B  (slow down animation if necessary) and the second click turns into a replace. The prev keys match the new keys (both length 2) but the old State is A and the new State is B. So it's a replace.
Changed to get old State from the prev state navigator in React state. The keys use prev from Reat state so so should the state.
Tried copying over the NavigationStack function but the zoom example didn't like having unique keys each navigaiton. A --> B then back to A then forward to C. Then use browser back fast (keep singer on CMD + back arrow) and the shared elements can drop. Think it's because C and B have different keys so get unmounted. When use same keys then they can reuse the existing scenes and the shared els are smooth. Think might apply this function back in NavigationStack - should work now that Scenes are unmounted as soon as native pop completes