Started app from deep link
If deep link when app closed (not running in background) then react context is null, even in MainActivity. Can't remember if it was always so. So removed onCreate of MainActivity because too early. Also, moved LinkActivity code into onCreate our of onDestroy otherwise the screen goes blank when LinkActiviyt is destoryed while it's waiting for MainActivity to start