Navigated up until crumbs match
Tracked crumb in intent. Then on new intent check if the crumbs changed and, if so, navigate up with the new intent.
A --> B --> C --> D Navigate back 2 first goes to C because all Activities have the same class. But C crumb is 2 and target crumb is 1 so it navigates up and reaches B!