Filtered out only the popped item
Don't need to remove items to the left to clear out popped items on Android now that one Activity is destoryed for each scene navigated back.
But this workaround was also used on iOS when clicking a tab to pop multiple scenes at once. That was because left it up to the animated popped scene to pop all the others so that it happened after the back navigation occured in React. But there's no need for that either. Just remove item when scene dealloc. There was a problem. Navigate A --> B --> C --> D and click tab. First pop removes B but then getDerived is called and the props are still the same because the pop hasn't happened in React yet. So it puts B back in! So changed getDerived to return null if data props not changed