Prevented unmount when not popped
On Android, the activitiy is destroyed for many reasons. e.g., when orientation changes. Mustn't unmount then. Must only unmount when actually popped. On Android, pops only happen from React code. So traked when popped and then checked true when popNative called so don't unmount if not popped from React. Only do this check on Android because  on iOS the pop can happen on client before the React pop. In this case want to unmount and then let the React navigation happen after that.
In short, on Android, only unmount scenes that have already been popped in React