Animated back to fluently opened
Fluently open scenes B and C and need the animations to work when navigating back. Can't get them from NavigationStack properties because B and C can have different animations. Need to read them from the Scene. The enter and exit on B, for example, refer to navigating back from B to A - the exit is B leaving and the enter is A returning. Need to set them on B even though one relates to A because A doesn't render during the fluent forward and that's when we need them!