Supported navigation stack in a modal
Mustn't finish the main activity when a fragment stack finiishes because it might be a nested stack in a modal. Instead, added a listener to when the fragment back stack hits 0 and then finished the main activity. This works because the default back press of react native is to do a popBackStackImmediate which sets it to 0.
Also had to give each fragement and backstack a unique tag/name otherwise can confilict. Could go back from 1 in nested to 0 in parent stack, but then it would go do 0 of the nested stack instead