Popped back on refresh
A --> B to A --> C then don't want scene B to be in the stack when navigating back (from C to A) because can see it animating. Popped when refreshing to remove it.
Unfortunately this meant removing the finish on empty back because refresh from A to C now pops A and gives an empty back stack. Need to find another way to end app when going back from first scene