Ensured elevation always increases
Navigation from A --> B to A --> C then index of C was 2 when it's added and then is changed to 1 when B is removed. But the elevatoin gets frozen at 2. So when navigate from C to D then both have elevation 2 and so C shows above (because of odd android replace behaviour). Created elvation by adding one to last elements elevation to guarantee it always increases