Removed support for dynamic add/remove
Decided to start simple. The normal usecase is to define the tabs upfront, not to move them around or add and remove later on. Keeping simple means don't need to setOffScreenPageLimit and don't need to remove fragment or set eleevation (elevation was for when newly added  flashed over the top of the current)