Added icon to tabs
This was incredibly tricky because couldn't find way to pass image as a prop. The ToolBarAndroid was the only component that did it and it was weird (plus ToolBarAndroid no longer in source?!?).
Copied the inlineImageLeft prop of TextInput which relies on it being a drawable resource. Fair compromise.
Made even trickier because tabs are recreated when new ones added or removed so have to update the icons if tabs change. Added a tabCount prop to TabBar to guarantee that onAfterUpdateTransaction runs - need to update from here otherwise to early and the new tabs won't exist