Removed the scenes from the stack's children
The stack's only child is the FragmentContainer. The scene's shouldn't be children because they're children of the FragmentContainer. Need to keep the child view hierarchy correct because things like the JSTouchDispatcher traverse it.
Need to return the scenes from the manager's children so that they're destroyed by React Native. When navigating back must return the removed scene from the manager so that React Native can remove it from its caches