Prevented bottom gap on orientation change
Change from portrait to landscape and scroll up then there was a gap at the bottom. Think the problem is the Coordinator Layout is doing the sizing and React Native doesn't know about it. So told React Native to leave sizing to the native side and seems to work?!
Still a problem with changing orientation to landscape then navigating to details- then can't scroll to the bottom!?