Got basic collapsing bar working
Copied code from react-native-collapsing-toolbar, especially the weird bits, and it started working. The weird bit is the setHeight via LayoutParams - without it it doesn't work. Had to add a couple of requestLayout & post(measureAndLayout) so it updates at start and after navigating back. Not sure why react-native-collapsing-toolbar didn't need them. Also didn't copy over the needsCustomLayoutForChildren on NavigationBarManager because it prevent the title view from updating on the details scene