Ensured coordinator scroll when collapsed content short
This took me about a week to work out! When the collpased content is shorter than the coordinator layout height then it doesn't scroll at all! That means some content can't be seen. Don't know how realistic this scenario is because the content should be long enough. In this case need to measure the gap between the collpased content and coordinator height and subract that from the measure spec. Add 2 to the coordinator height to ensure that the collapsed content clears the height