Displayed search with coordinator layout
Can't render the search inside the coordinator layout because android doesn't like it. Need to render it as a sibling of the coordinator layout so that the absolute position and size work.
There was a problem: tap search, tap color, tap back, close search then scrolling up disn't scroll the content up. It was because the keyboard wasn't closed properly when navigating?! Had to request focus and rediisplay keyboard otherwise