Changed can navigate check to go against fragment manager
The last check didn't work when tried in the zoom sample. Threw the same error after double tap R and then Android Home. Maybe the tiiming was different but it went into the onAttachedToWindow of the new Stack which ran the navigation because canNavigate was true.
Using a better check which is whether the FragmentManager allows the navigation. Still re-run navigation onResume