Populated tabs from tab views
Thought didn't need this - remember deleting it. But turns out that sometimes the onAttachedToWindow of the TabBarView runs before the TabView is added as a child - so the TabBarView can't find the TabView. Sure it's only if it's bottom tabs so added check. Plus can't run it if top tabs because it adds it twice when there's a TabBar in the NavigationBar