Future-proofed modal hand back on ios
On ios, the onRequestClose can be called when showing modal with pageSheet and dismissed with a swipe. Thought this fix was in already, https://github.com/facebook/react-native/commit/8e5fac89bbdcc3028bb5d81a358969a235abf991, but it's not working. Adding the check so it skips on ios when the fix lands