Returned tabbarviewitem from adapter
Tried returning the first child but goes wrong when adding new items. It calls instantiateitem for the new item but it already has a parent of the tabbarviewitem. Got away with this before because the TabFragment removed the parent at the right time?
It's better using the tabbarviewitem but tried before and it broke the coordinatorlayout. Scrolling up left a gap at the bottom. Fixed by changing the base class to LinearLayout instead of ViewGroup