Matched ios13 override behaviour to ios12
If the prop can be set in React then always take the React value. If it can't then don't overwrite it.
It's complicated to use the default for React props. Could do it for navigation bar but haven't done it elsewhere. Need to get the default value in didMoveToWindow because it's not set early enough to use in the set accessors