Ensured tab loaded before setting selected tab
When first loading the selectedTab prop is set before any tabs are loaded. So added set the selected tab when addedToWindow.
But can also add a new tab and set it to be selected and then face the same problem. So added a data set observer and set the selecfed tab in there. Checked again if tabs loaded because might be adding two tabs at once and selecting the last