Put in optimistic locking for tab change
Two tabs A and B. Change quickly from A to B then back again. The first change happens and is sent to js. then the second change happens before the first returns. When the first change returns from js it sets the tab back to B and sends the change to js. Then the second change returns and sets back to A and sends change to js... this goes on forever and causes tabs to flicker between the two.
This strategy is in elsewhere (NavigationStack and SearchView). It's needed whenever change happens on both native and js. Ensures the js ones are thrown away if there's been a subsequent changes on native. The js will catch up with native in the end