Excluded data set change from native event count
If tabs A, C with C selected and insert tab B between them then this changes the selectedTab from 1 to 2. But this isn't a user triggered event - it changed because a tab was inserted - so don't want to increment the native count. For example, might want to set C selected at the same time as inserting it - if increment the native cound then the tab change is ingnored.
This brings it inline with ios where the tab change takes precendence