Moved willNavigateBack event into navigation stack
Need to start and suspend navigation so can't do it in Scene. Needs to be aware of the whole stack.  Could try and put it in scene but it would have to go in the from scene - might be complicated when double tapping tab?! Doesn't add anything to put it in scene anyway, because the double tap problem doesn't exist when the navigation happens on the client first