Dropped handler dependency
Tried to create failing test when dropping the handler dependency but couldn't!! Whenever the navigationEvent changes the useEffect reruns and picks up the latest handler.
Would only need the handler depedency if there was some listener running because it fires independent of navigationEvent changing. This is the usual usecase for useEffect which is why was surprised it's not needed