Prevented calling navigated hook when peeking on ios
If fluently navigate from A to A -> B -> C then when peeking back to B the navigated fired when the swipe started. It fired again when the swipe ended. It shouldn't fired when at the start - it only did because of the temporary navigationEvent created by Scene so that it renderes.
Set a private 'peek' indicator in context to identify this situation so that the navigated hook can ignore this temp event.
Also ignored  onBeforeNavigate if android because it's only on iOS that navigation's happen on client first