Updated to latest rollup and typescript
Wasn't easy because had to swap out the rollup typescript package, too. Old package wasn't supported and couldn't get it working with latest typescript ( or maybe latest rollup?).
The new rollup typescript is very different and much slower. It's slower because it does full type checking. This is better and it found a few bugs.
Had to add an include option otherwise it wen through all files - the old one jwas better because it used the input and went from there.
The tests are more than 6x slower but it's not bad if comment out all but one test suite. So that's bearable