Went back to the FragmentPagerAdapter
It's better because not all tabs have to be attached to window (in view) - only the current one + one other. So with 4 tabs only 2 are attached. Couldn't get this to work before so there was no point using the fragment adapter. But ViewPager2 made me take another look because thought things would be easier now - but ViewPager2 is final so can't be subclassed so had another go with fragment adapter. Basic testing suggests it could work - things have changed since first tried, in particular the overridden requestLayout which might make the difference.
With 3 or more tabs of NavigationStacks, when changing tab go a black background shown during the animation. The NavigationStackViews only navigate once they're added to window.  So the stacks that aren't added (other tabs) stay blank - black background. This is what shows when changing tab first time. Inside TabFragment forced a navigation by manually calling onAfterUpdateTransaction