Survived an app restart after, e.g., changing language
In #409, for example, changing language causes app to restart (native not js). Android tries to recreate the fragments but can't find a default constructor so falls over. Adding empty constructors isn't enough because then it can't find a view if return null from onCreateView. So returned a dummy empty view