Removed fragments when view dropped
A -> B and B has tabs then the tab fragments should go when navigating back to A. Copied NavigationStack approach and removed the parent fragment when view dropped.
Needed to fix up the child view methods in TabBarItemManager otherwise react native can't find the views to drop. In Twitter the TabBarPager is a descendant of TabBarItem so it has to return correct children to find it