Started navigation as soon as child scenes available
When starting a new stack the initial scene fragment wasn't loaded until the view is attached to window. This causes a black background to show when switching tabs with stacks. Instead, ran the first navigation as soon as the scenes are added.
Catered for when the stack is loaded without a navigation. In this case don't need a start navigation because the onUpdateAfterTransaction fires after children and keys set when the navigation finally happens. Identified this case by passing empty keys. Only need special handling for when it loads with a navigation - when keys not empty