Ensured large title starts expanded on iOS 14
Not sure what's wrong with RN on iOS 14. If set contentOffset={{x: 0, y: -1}} on the Container ScrollView then sometimes it works and sometimes it doesn't!? I put breakpoint in setContentOffset and it doesn't call this in a predictable way - one way it worked and another it didn't.
This workaround hides the content and then immediately shows it in a setTimeoute of 0 millis. Hiding it for first render means the large titles start expanded. Then showing it immediately works