Got crumb from scene key
The indexOf doesn't work in ios14. In Twitter sample navigate to tweet A then tweet B and then tap the Home tab to navigate back to the start. It goes back to start and then immediately shows Tweet A again! That's because when the willShow is called the viewControllers aren't in the corret order - the Home scene is at index 1 instead of 0!! So it goes back to 0 (built in behaviour) then navigate to 1