Hid the navigation bar in viewWillLayoutSubviews
It's an iOS bug where the navigation bar doesn't hide when coming from search results, see https://stackoverflow.com/questions/8950541/setnavigationbarhiddenyes-doesnt-work-with-the-searchdisplaycontroller
Followed the SO answer and used viewWillLayoutSubviews. Kept the hide in the viewWillAppear as well otherwise saw a weird animation glitch in the zoom example where a line appears where the navigation bar would've been.