Decided where to hide based on whether search open
Noticed a bug that never seen before on iOS 12. In zoom example, tap in search box and select a color then go back and select color again and the navigation bar appears in the search results - can just see it as the search results animate offscreen.
Turns out when the search controller is open the navigationbar is hidden! iOS hides it!! So when show it in viewWillAppear of details then it grid header shows in the search results!!
The trick is to hide the navigation bar either in viewWillAppear if the search controller is closed and in viewWillLayoutSubviews if it's open.