Swizzled the controllerForStyle so it reaches the scene
Turning on UIViewControllerBasedStatusBarAppearance didn't work. IOS asks the root view controller for it's preferredStyle - in react native the root view controller is a UIViewController and so it never reaches the SceneViewController.
Method swizzling means can swap the UIViewController controllerForStyle implementation for another. Changed it to look for the NavigationStack and then the top scene within the stack. This gets it to the SceneController which is then asked for its preferred style. Got the idea from react native screens https://github.com/software-mansion/react-native-screens/commit/ddf1f81d5e32c9cf76179cc9396f86a0880ba0cc.