Rendered status bar always
Was having lots of problems with the android twitter tabs example. What if don't have status bar on home tab but do on notifications tab - these tabs are all in a single scene so can't handle it at the scene fragment level. Was trying to handle it at the tab fragment level but then when navigated from notifications to tweet and back again the scene fragment overrode the tab fragment.
Makes much more sense to always have a statusbar along with the. navigation bar. Then it all just works. Don't need to reset because the status bar will always set. Guess there may be cases when there's no navigation bar on android?! but can just add one with hidden true