Used the presenting preferences
Have a view controller with a light status bar, for example. Then open a search then the UISearchController becomes viewcontroller for the statusbar. But there's no scene or status bar for that viewcontroller so the status bar settings go back to the defaults.
So swizzle the preference methods so that can redirect it back to the scene - the presenting view controller.
These swizzles don't get in the way if it's a scene because the scene overrides these methods which ignores the swizzles