Simplified swizzle with concrete nav and search controllers
Implemented custom navigation controller so can return visibleViewController from the child view for status requests. Simplifies the swizzle because only need to return a navigation controller - not the visible view controller.
Similarly for the search view controller. Don't need to swizzle the preferences because can satisfy them in the dedicated search controller