Removed outdated "Thinking in Navigation"
This was a reply to "Thining in React", but there's better ways to think in navigation now. Also, hard to bring the doc up to date to latest navigation when it needs to look like the original React doc which still uses class components. Would need to useEffect for the timeout now, for example