Dropped hash param from navigate and refresh
The html5 history api doesn't respect the hash - it doesn't scroll to the fragment identifier. The navigation router should only build 'good' urls - so don't want to append the hash when programmatically navigating. The only time a hash is relevant is for links. If click a link still want to keep it in the url when programmatic because confusing if hash only there if shift-click the url, for example.