Fixed unloaded hook tests after sync flush
Some unloaded hook tests failed after switching to flushSync in NavigationHandler. Take unloaded hook test A --> B for example.

The problem was that the navigation happens outside React, the onNavigate event fires, the NavigationHandler receives the event and does a flushSync set state. Then the unloaded hook useEffect on A immediately removes the onNavigate listener and adds it again because sync flush. But that handler isn't called because we're already inside the onNavigate loop through handlers.
This problem doesn't happen if the navigation is initiated from within React. It only if navigation happens outside so that the set state is done when looping through onNavigate handlers. If the useEffect is removed then added either before or after the onNavigate then the unloaded hook works fine.
So added a new afterNavigate even that fires after  the onNavigate, at the very end of the navigation after the browser history is updated. Then listened for this in NavigationHandler and flushed sync then. That way the onNavigate handlers are all called, in particular the unloaded hook.