Prevented clicking when dragging
Tested with button in bottom sheet. Dragged sheet by holding button and if the sheet settled back in same detent (collapsed to collapsed) then the button click fired. Same but as swiping with tabbar on android - so added a dragging field that follows same approach as swiping field in tabbar