Always ran custom animations, even for shared els
Material motion recommend different transitions when running shared elements - like MaterialElevateScale to fades and scales the exiting scene. Don't want to use transitions (Enter, Exit, Return and Reenter on Fragment) because it's confusing to have transitions and animations. Transitions don't seem as flexible for scenes (how to do a basic scale?!) so can't get rid of animations. Also, Android doesn't wait for the transitions to complete before calling onDestroy!!! This means the leaving scene disappears as it's immediately unmounted on React side.
So allowed user to write their own 'MaterialElevateScale' as an animation and use it for the shared el transition