Loaded shared els same when returning as going
With fragment A and B, the shared element enter and return transition need to be set on B! But the enter transition needs to be postponed on B when entering and A when returning! So passed 2 fragments into the transitioner.
Means can let the shared element PreDraw listener take care of loading when returning (same as when entering)