Removed drawing view id prop for simplicity
The MaterialContainerTransform expects the id to be a resource?! The React Native ids aren't resources so android throws error when it can't find it. Could get round it by overriding getResources of SharedElementView, providing custom Resources and catching error from getResourceName. Not sure of consequences.
Also, it didn't work that well - the scrim was only applied over the drawing view but the from shared element disappeared completely at start of animation and then came back (not sure why - set duration to 5000 so could see it).
Also not sure of the timings. Maybe setting the drawingViewId from findNodeHandle will happen too late? Can't make it part of the postpone enter transition.
All in all, simplest to remove for now.
See https://github.com/material-components/material-components-android/issues/1224 for why drawing view id is helpful - this example prevents the scrim appearing over the top of the toolbar.