Supported backgroundColor prop
Using the style prop of the FloatingActionButton. A bit inconsistent with other navigation components (barbutton etc.) but might make sense for FABs. FABs are standalone (bar button gets tint from navigation bar, for example) and want to be able to animate, e..g.,  opacity and transform so gonna need to use the style prop anyway